+++ packages/client/src/utils/validate_id.ts
